examples/deployments/etl-pipeline/requirements.txt
# Daita Agents Framework
daita-agents==0.9.0

# LLM provider
openai>=1.0.0

# Database
asyncpg>=0.29.0

# Data processing
pandas>=1.5.0
numpy>=1.20.0

# Utilities
python-dotenv>=1.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
